requirements.txt
google-generativeai==0.8.5
requests==2.32.5
python-dotenv==1.2.1
